+++ app/features/workout/card/expanded/title/EditWorkoutTitleExpandedActions.js
@@ -1,7 +1,6 @@
 import {
     COLLAPSE_WORKOUT_SET,
     PRESENT_WORKOUT_EXERCISE,
-    PRESENT_WORKOUT_VIDEO_PLAYER,
 } from 'app/ActionTypes';
 import * as Analytics from 'app/services/Analytics';
 import * as SetsActionCreators from 'app/redux/shared_actions/SetsActionCreators';
@@ -29,30 +28,10 @@ export const presentExercise = (setID, exercise, bias) => (dispatch, getState) =
     });
 };
 
-export const presentWatchVideo = (setID, videoFileURL) => (dispatch, getState) => {
-    const state = getState();
-    Analytics.setCurrentScreen('workout_watch_video');
-    logWatchVideoAnalytics(setID, state);
-    
-    dispatch({
-        type: PRESENT_WORKOUT_VIDEO_PLAYER,
-        setID: setID,
-        videoFileURL: videoFileURL
-    });
-};
-
 const logEditExerciseNameAnalytics = (setID, exercise, state) => {
     let is_working_set = SetsSelectors.getIsWorkingSet(state, setID);
 
     Analytics.logEventWithAppState('edit_exercise_name', {
-        is_working_set: is_working_set
-    }, state);
-};
-
-const logWatchVideoAnalytics = (setID, state) => {
-    const is_working_set = SetsSelectors.getIsWorkingSet(state, setID);
-    
-    Analytics.logEventWithAppState('watch_video', {
         is_working_set: is_working_set
     }, state);
 };
