+++ app/features/workout/card/expanded/title/EditWorkoutTitleExpandedScreen.js
@@ -7,8 +7,8 @@ import * as Actions from './EditWorkoutTitleExpandedActions';
 const mapDispatchToProps = (dispatch) => {
     return bindActionCreators({
         saveSet: Actions.saveSet,
-        tapExercise: Actions.presentExercise,
-        tapCollapse: Actions.tapCollapse,
+        tappedExercise: Actions.presentExercise,
+        tappedCollapse: Actions.tapCollapse,
     }, dispatch);
 };
